==> All_Other_Analysis/Scripts/01_Soil_chemistry.Rmd <==
---
title: "Experiment BX-feedback M18W19  <br> Soil parameters"
author: "Valentin Gfeller"
date: "`r Sys.Date()`"
output: 
  html_document: 
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
#global knitter option
knitr::opts_chunk$set(fig.align = "center", dev = 'svglite', echo = FALSE)

# load packages
library(ggbeeswarm); packageVersion("ggbeeswarm")
library(ggforce); packageVersion("ggforce")
library(ggpubr); packageVersion("ggpubr")
library(viridis); packageVersion("viridis")
library(emmeans); packageVersion("emmeans")
library(mgcv); packageVersion("mgcv")
library(here); packageVersion("here")
library(FactoMineR); packageVersion("FactoMineR")
library(factoextra); packageVersion("factoextra")
library(car); packageVersion("car")
library(knitr); packageVersion("knitr")
library(tidyverse); packageVersion("tidyverse")

# source handy functions
source(here("Functions", "fun_vg.R"))

# set global option for ggplot2
theme_set(theme_bw(base_size = 12))

# renv::init()
# renv::snapshot()
```


```{r, include=FALSE}
#Import data
d.chem_chang <- read_delim(file = here("Data","01_Soil",
                                       "Exp_m18w19_soil_LBU_GIUB.csv"),
                           delim = ",", col_names = TRUE) %>%
                           mutate(trt = factor(trt, levels = c("W", "b")))

d.chem_chang %>% str

# Exclude the following variables due to lack of variation among samples: 
# * clay
# * silt
# * sand
# * humus

d.chem_chang_red <- d.chem_chang %>% 
  select(!c(clay, silt, sand, humus))

#wide to long data frame
d.chem_chang_long <- d.chem_chang_red %>% 
  pivot_longer(7:24, names_to = "compound", values_to = "quantity")

```

# Overview field
```{r, echo=FALSE, fig.height=4.5, fig.width=8.5}
## check position on field
p <- d.chem_chang_red %>% 
  mutate(width = factor(width, levels = c("3", "2", "1")), 
         length = factor(length)) %>% 
  ggplot(aes(length, width, colour = trt, fill = trt)) +
  geom_point(shape = 22, size = 25, alpha = 0.4) +
  geom_text(aes(label = plot_id ), color = "black",
            hjust = 0.5, vjust = 0.5, show.legend = FALSE)+
  scale_color_manual(name = "Soil conditioning", 
                     labels = c(W = "WT", b = "bx1 (W22)"), 
                                values = c(W = "gold2", b = "darkgreen"))+
  scale_fill_manual(name = "Soil conditioning", 
                     labels = c(W = "WT", b = "bx1 (W22)"), 
                                values = c(W = "gold2", b = "darkgreen")) +
  theme_classic()

p

```

# Field gradients
## Have a look at one of the soil parameters to get an overview
```{r, out.width="100%", fig.height=20, message=FALSE, warning=FALSE, eval=F}
p.length <- d.chem_chang_long %>% 
  filter(compound == "potassium_CO2") %>% 
  ggplot(aes(x = length, y = quantity)) +
  geom_point(size = 1.5, alpha = 3/4) +
  geom_text(aes(label = plot_id), alpha = 1/3) +
  stat_smooth(method = "lm", se = FALSE, formula = "y ~ x") +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~"))) + 
  facet_wrap(vars(compound), scales = "free", ncol = 2)

p.length

```

```{r, out.width="100%", fig.height=20, message=FALSE, error=FALSE, warning=FALSE, eval=F}
p.width <- d.chem_chang_long %>% 
  filter(compound == "potassium_CO2") %>% 
  ggplot(aes(x = width, y = quantity)) +
  geom_point(size = 1.5, alpha = 3/4) +
  geom_text(aes(label = plot_id), alpha = 1/3) +
  stat_smooth(method = "lm",se = FALSE, formula = "y ~ x") +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~"))) + 
  facet_wrap(vars(compound), scales = "free", ncol = 2)

p.width
```

```{r, fig.height=4.5, fig.width=8.5, error=FALSE, message=FALSE, eval=F}
d.chem_chang %>% 
  mutate(width = factor(width, levels = c("3", "2", "1")), 
         length = factor(length)) %>% 
  ggplot(aes(length, width, fill = potassium_CO2)) +
  geom_point(shape = 22, size = 20) +
  scale_fill_viridis(discrete = FALSE, option = "D", direction = -1) +
  geom_text(aes(label = plot_id),
            hjust = 0.5, vjust = -2.75, show.legend = FALSE) 

# lm
m.pot <- d.chem_chang %>% 
  lm(potassium_CO2 ~ (trt + length + width)^2, data = .)

Anova(m.pot) %>% broom::tidy() %>%  mutate(across(where(is.numeric), ~ as.character(signif(., 3)))) %>% kable()
broom::glance(m.pot)[1, 1:6] %>%  mutate(across(where(is.numeric), ~ as.character(signif(., 3)))) %>% kable()
```


# Multivariate analysis
```{r, include=FALSE}
# set row names
d.chem_chang_red <- d.chem_chang_red %>% as.data.frame()
rownames(d.chem_chang_red) <- d.chem_chang_red$plot_id

# complete missing values
library(missMDA)
d.chem_chang_red %>% str
d.imputePCA <- d.chem_chang_red[, 7:24]
nPCs <- estim_ncpPCA(d.imputePCA, ncp.min = 0, ncp.max = 5) 
d.chem_chang_completed <- imputePCA(d.imputePCA, ncp = nPCs$ncp, scale = TRUE)
d.chem_chang_completed <- d.chem_chang_completed$completeObs %>% as.data.frame()
d.chem_pca <- bind_cols(d.chem_chang_red[, 1:6], d.chem_chang_completed)

# Compute PCA
tibble(col_names = colnames(d.chem_pca), 
       col_num = c(1:length(colnames(d.chem_pca)))) %>%
  print(n = 100)

d.pca.soil <- d.chem_pca[, c(3:5, 7:24)] %>% 
  mutate(trt = factor(trt, levels = c("W", "b")))

nutri.pca.FR <- PCA(d.pca.soil,
                 ncp = 5,
                 scale.unit = TRUE,
                 quali.sup = 1,
                 quanti.sup = c(2, 3),
                 graph = FALSE)


#Check PCA
summary(nutri.pca.FR)
#str(nutri.pca.FR)
```

# #Fig.2A
```{r}
set.seed(10)
fviz_pca_biplot(nutri.pca.FR, axes = c(1, 4), repel = TRUE,
                col.quanti.sup = "darkblue", geom = c("point"),
                col.var = "black",  title = NULL, pointsize = 2.5,
                alpha.var = 0.7) +
  labs(x = paste0("Soil chemistry PC1 (", 
                  nutri.pca.FR$eig["comp 1", "percentage of variance"] %>% 
         round(digits = 1), "%)"),
       y = paste0("Soil chemistry PC4 (", 
                  nutri.pca.FR$eig["comp 4", "percentage of variance"] %>% 
         round(digits = 1), "%)")) +
  theme_bw(base_size = 14)

ggsave(here("Figures", "individual_figures", "Fig.2A.svg"), 
       width = 20, height =  13, units = "cm")
```

```{r, include=FALSE}
d.ext <- nutri.pca.FR$ind$coord %>% as_tibble(rownames = "plot_id")

d.chem_pca_dim <- left_join(d.chem_pca, d.ext)

d.chem_pca_dim_save <- d.chem_pca_dim %>% 
  rename(c("PCA_soil_chem_1" = "Dim.1", "PCA_soil_chem_2" = "Dim.2", "PCA_soil_chem_3" = "Dim.3", "PCA_soil_chem_4" = "Dim.4", "PCA_soil_chem_5" = "Dim.5")) %>% 
  select(1:5, starts_with("PCA"))

write.csv(d.chem_pca_dim_save, row.names = FALSE, 
          file = here("Data", "01_Soil", "Exp_m18w19_PCA_soil_chem.csv"))
```

```{r}
d.chem_pca_dim <- read_delim(file = here("Data","01_Soil",
                                         "Exp_m18w19_PCA_soil_chem.csv"),
                           delim = ",", col_names = TRUE) %>%
                           mutate(trt = factor(trt, levels = c("W", "b")))
str(d.chem_pca_dim)
```

# #Fig.S3 (top)
```{r, message=FALSE, error=FALSE, fig.height=3.5, fig.width=12}
fac_lab <- c(PCA_soil_chem_1 = "PC1", PCA_soil_chem_2 = "PC2", 
             PCA_soil_chem_3 = "PC3", PCA_soil_chem_4 = "PC4",
             PCA_soil_chem_5 = "PC5")

# convert to long format
d.chem_pca_dim_long_pos <- d.chem_pca_dim %>%
  select(1:5, starts_with("PCA")) %>% 
  pivot_longer(6:10, names_to = "compound", values_to = "dim_coord")

# position latitudinal
d.chem_pca_dim_long_pos %>% 
  ggplot(aes(x = length, y = dim_coord, color = trt)) +
  geom_point(size = 2, alpha = 3/4) +
  scale_x_continuous(breaks = seq(1, 7, 1), limits = c(1, 7)) +
  theme_bw(base_size = 12) +
  stat_smooth(method = "lm",se = FALSE, fullrange = TRUE, colour = "dimgrey",
              alpha = 3/4) +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~"),
               colour="dimgrey", fontface="bold")) +
  scale_color_manual(name = "Soil conditioning", 
                    labels = c(W = "W22", b = "bx1"), 
                    values = c(W = "gold2", b = "darkgreen")) +
  xlab("Field position: length") +
  ylab("Coordinates of dimension")+
  facet_wrap(vars(compound), scales = "free", ncol = 5,
             labeller = labeller(compound = fac_lab)) +
  theme(legend.position = "none")


ggsave(here("Figures", "individual_figures", "Fig.S3_top.svg"), 
       width = 32, height =  8, units = "cm")
```

# #Fig.S3 (bottom)
```{r, message=FALSE, error=FALSE, fig.height=3.5, fig.width=12}
# convert to long format
d.chem_pca_dim_long_pos <- d.chem_pca_dim %>%
  select(1:5, starts_with("PCA")) %>% 
  pivot_longer(6:10, names_to = "compound", values_to = "dim_coord")

# position longitudinal
d.chem_pca_dim_long_pos %>% 
  ggplot(aes(x = width, y = dim_coord, color = trt)) +
  geom_point(size = 2, alpha = 3/4) +
#  geom_text(aes(label = plot_id), alpha = 1/3) +
  scale_x_continuous(breaks = seq(1, 3, 1), limits = c(1, 3)) +
  theme_bw(base_size = 12) +
  stat_smooth(method = "lm",se = FALSE, fullrange = TRUE, colour = "dimgrey",
              alpha = 3/4) +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~"),
               colour="dimgrey", fontface="bold")) +
  scale_color_manual(name = "Soil conditioning", 
                    labels = c(W = "W22", b = "bx1"), 
                    values = c(W = "gold2", b = "darkgreen")) +
  xlab("Field position: width") +
  ylab("Coordinates of dimension")+
  facet_wrap(vars(compound), scales = "free", ncol = 5,
             labeller = labeller(compound = fac_lab)) +
  theme(legend.position = "none")


ggsave(here("Figures", "individual_figures", "Fig.S3_bottom.svg"), 
       width = 32, height =  8, units = "cm")
```

# #Fig.S4
```{r}
d.f <- d.chem_chang_red %>% 
  mutate(width = factor(width, levels = c("3", "2", "1")), 
         length = factor(length)) 

plot_chem_2D <- function(variable) d.f %>% 
  ggplot(aes(length, width, fill = {{variable}})) +
  geom_point(shape = 22, size = 20) +
  scale_fill_viridis(discrete = FALSE, option = "D", direction = -1,
                     na.value = "lightgrey") +
  geom_text(aes(label = plot_id),
            hjust = 0.5, vjust = -2.75, show.legend = FALSE) +
  xlab("Field position: length") +
  ylab("Field position: width") +
  theme(legend.position = "top", legend.key.width = unit(2, "cm"),
        text = element_text(size = 14),
        legend.margin = margin(b =  0, unit='cm'),
        legend.box.margin = margin(b = -10, unit = "pt"))

d.chem_chang_red %>% colnames()
#Plots
p.pH <- plot_chem_2D(pH) %>% plot()
p.phosphorus_CO2 <- plot_chem_2D(phosphorus_CO2)%>% plot()
p.potassium_CO2 <- plot_chem_2D(potassium_CO2)%>% plot()
p.magnesium_CO2 <- plot_chem_2D(magnesium_CO2)%>% plot()
p.nitrate_H2O <- plot_chem_2D(nitrate_H2O)%>% plot()
p.phosphorus_H2O <- plot_chem_2D(phosphorus_H2O)%>% plot()
p.potassium_H2O <- plot_chem_2D(potassium_H2O)%>% plot()
p.calcium_H2O <- plot_chem_2D(calcium_H2O)%>% plot()
p.magnesium_H2O <- plot_chem_2D(magnesium_H2O)%>% plot()
p.iron_H2O <- plot_chem_2D(iron_H2O)%>% plot()
p.phosphorus_AAE <- plot_chem_2D(phosphorus_AAE)%>% plot()
p.potassium_AAE <- plot_chem_2D(potassium_AAE)%>% plot()
p.magnesium_AAE <- plot_chem_2D(magnesium_AAE)%>% plot()
p.boron_AAE <- plot_chem_2D(boron_AAE)%>% plot()
p.copper_AAE <- plot_chem_2D(copper_AAE)%>% plot()
p.iron_AAE <- plot_chem_2D(iron_HNO3)%>% plot()

ggarrange(p.pH,
          p.phosphorus_CO2,
          p.potassium_CO2,
          p.magnesium_CO2,
          p.nitrate_H2O,
          p.phosphorus_H2O,
          p.potassium_H2O,
          p.calcium_H2O,
          p.magnesium_H2O,
          p.iron_H2O,
          p.phosphorus_AAE,
          p.potassium_AAE,
          p.magnesium_AAE,
          p.boron_AAE,
          p.copper_AAE,
          p.iron_AAE,
          ncol = 3,
          nrow = 6)

ggsave(here("Figures", "individual_figures", "Fig.S4.svg"), 
       width = 50, height =  65, units = "cm")
```

# ##Fig.2B
```{r}
d.chem_pca_dim_f <- mutate(d.chem_pca_dim, width = factor(width, levels = c("3", "2", "1")), length = factor(length))

d.chem_pca_dim_f %>% 
  ggplot(aes(length, width, fill = PCA_soil_chem_1)) +
  geom_point(shape = 22, size = 20) +
  scale_fill_viridis(discrete = FALSE, option = "D", direction = -1,
                     name = "Soil chemistry PC1") +
  geom_text(aes(label = plot_id),
            hjust = 0.5, vjust = -2.75, show.legend = FALSE) +
  xlab("Field position: length") +
  ylab("Field position: width")  +
  theme_bw(base_size = 14)+
  theme(legend.position = "top", legend.key.height = unit(.3, "cm"),
        legend.margin = margin(b =  0, unit='cm'),
        legend.box.margin = margin(b = -10, unit = "pt"))

ggsave(here("Figures", "individual_figures", "Fig.2B.svg"), 
       width = 20, height =  13, units = "cm")
```

# #Fig.5A
```{r}
d.chem_pca_dim_f %>% 
  ggplot(aes(length, width, fill = PCA_soil_chem_1)) +
  geom_point(shape = 22, size = 20) +
  scale_fill_viridis(discrete = FALSE, option = "D", direction = -1,
                     name = "Soil chemistry PC1") +
  geom_text(aes(label = plot_id),
            hjust = 0.5, vjust = -2.5, show.legend = FALSE,
            size = 5) +
  xlab("Field position: length") +
  ylab("Field position: width")  +
  theme_bw(base_size = 18)+
  theme(legend.position = "top", legend.justification = "left",
        legend.key.height = unit(.3, "cm"),
        legend.margin = margin(b =  0, unit='cm'),
        legend.box.margin = margin(b = -10, unit = "pt"))

ggsave(here("Figures", "individual_figures", "Fig.5A.svg"), 
       width = 20, height =  13, units = "cm")
```


# Version control
```{r}
sessionInfo()
```
