==> All_Other_Analysis/Scripts/02_BXs_field.Rmd <==
---
title: "Experiment BX-feedback M18W19  <br> Benzoxazinoid concentrations in the soil at maize harvest, wheat sowing, and wheat growth"
author: "Valentin Gfeller"
date: "`r Sys.Date()`"
output: 
  html_document: 
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
#global knitter option
knitr::opts_chunk$set(
	echo = FALSE,
	fig.align = "center",
	warning = FALSE,
	dev = "svglite",
	fig.width = 6,
	fig.height = 5
)
# load packages
library(tidyverse); packageVersion("tidyverse")
library(ggbeeswarm); packageVersion("ggbeeswarm")
library(ggpubr); packageVersion("ggpubr")
library(viridis); packageVersion("viridis")
library(car); packageVersion("car")
library(emmeans); packageVersion("emmeans")
library(mgcv); packageVersion("mgcv")
library(here); packageVersion("here")
library(FactoMineR); packageVersion("FactoMineR")
library(factoextra); packageVersion("factoextra")
library(purrr); packageVersion("purrr")
library(tidytext); packageVersion("tidytext")
library(missMDA); packageVersion("missMDA")
library(cowplot); packageVersion("cowplot")
library(knitr); packageVersion("knitr")

# source handy functions
source(here("Functions", "fun_vg.R"))
source(here("Functions", "fun_vg_more.R"))

# set global option for ggplot2
theme_set(theme_bw(base_size = 12))

```


```{r, include=FALSE}
#Import data
d.bx_long <- read_delim(file = here("Data", "01_Soil", "Exp_m18w19_soil_BX.csv"),
                        delim = ",", col_names = TRUE) %>%
  mutate(trt = factor(trt, levels = c("W", "b")))

d.soil_chem <- read.csv(here("Data", "01_Soil", "Exp_m18w19_PCA_soil_chem.csv"), 
                        sep = ",", header = T) %>% 
  select(width, length, starts_with("PCA"))


# append PCA of soil analysis
d.bx_long <- left_join(d.bx_long, d.soil_chem, by =c("length", "width"))%>% 
  mutate(Compound_name = factor(Compound_name, 
                                levels = c("DIMBOA-Glc", "HDMBOA-Glc",  
                                           "DIMBOA","HMBOA",
                                           "MBOA", "AMPO")),
         sample_origin = factor(sample_origin, 
                                levels = c("Conditioning_phase", "Wheat_sowing",
                                           "Wheat_end_veg_stage")))
```

# End conditioning phase (at amize harvest)
# #Fig.2C
```{r include=FALSE}
d.bx_red <- d.bx_long %>% 
  filter(sample_origin == "Conditioning_phase")

d.bx_red %>%
  group_by(Compound_name) %>%
  summarise(
    n = sum(!is.na(Concentration))
  )
```

```{r}
# test for differences between genotypes by Wilcoxon Rank Sum test and correct for multiple testing
p_values <- d.bx_red %>% 
  mutate(Concentration = case_when(
    is.na(Concentration) ~ 0,
    TRUE ~ Concentration))%>%
  group_by(Compound_name) %>% 
  nest() %>% 
  mutate( fit = map(data, ~ wilcox.test(Concentration ~ trt, data = ., exact = FALSE)),
          results = map(fit, broom::glance)) %>% 
  unnest(results)%>% 
  select(1, 5)%>%
  ungroup() %>% 
  mutate(
    n      = length(.$p.value),
    p.ad   = p.adjust(p.value, method = "fdr"),
    p.ad_c = case_when(
      p.ad < 0.001 ~ as.character("p < 0.001"),
      TRUE  ~ paste0("p = ", sprintf("%.3f", p.ad))),
    x      = 1.5,
    y      = Inf)

# check values above LOD
t.lod <- d.bx_red %>%
  group_by(Compound_name, trt) %>%
  summarise(
    n        = n(),
    Above_LOD = sum(Concentration != 0, na.rm = TRUE),
    .groups = "drop") %>%
  kable(caption = "Count values above limit of detection")

# use information for plot
d.lod <- tibble(
  label = c("LOD"),
  Compound_name = c("HDMBOA-Glc", "DIMBOA", "HMBOA", "MBOA", "AMPO"),
  trt = "b",
  y = 0) %>% 
  mutate(Compound_name = factor(Compound_name, 
                                levels = c("DIMBOA-Glc","HDMBOA-Glc", "DIMBOA",
                                          "HMBOA", "MBOA", "AMPO")))


d.bx_red %>% 
  ggplot(aes(x = trt, y = Concentration))+
  stat_summary(aes(fill = trt), fun = mean, geom = "bar", 
               position = position_dodge(width = 0.8),
               color = "black",  size = 0.7, alpha = 0.85, show.legend = FALSE) +
  stat_summary(fun.data = mean_se, geom = "errorbar", 
               position = position_dodge(width = 0.8),
               size = 0.8, color = "black", width = 0.2, alpha = 0.65) +
  geom_quasirandom(aes(color = PCA_soil_chem_1), size = 1.8, shape = 16, alpha = 1,
                     width = 0.2) +
  ylab("Concentration (ng/ml)") +
  xlab("Maize genotype") +
  scale_y_continuous(expand = expansion(mult = c(0, 0.17))) +
  scale_x_discrete( labels = c(W = "WT", b = expression(italic(bx1)))) +
  scale_color_continuous(name = "Soil chemistry PC1") +
  scale_fill_manual(values = c(W = "gold2", b = "darkgreen")) +
  facet_wrap(vars(Compound_name), nrow = 1, scales = "free") + 
  geom_text(data = p_values, 
            mapping = aes(x = x, y = y, label = p.ad_c, fill = NULL),
            vjust = 2, size = 3.5) +
  geom_text(data = d.lod,
      mapping = aes(x = trt, y = y, label = label, fill = NULL),
      vjust = -1, size = 3) +
  theme(legend.position = "top", legend.key.height = unit(.3, "cm"),
        legend.margin = margin(b =  0, unit='cm'),
        legend.box.margin = margin(b = -10, unit = "pt"))

ggsave(here("Figures", "individual_figures", "Fig.2C.svg"), 
       width = 30, height =  8, units = "cm")

```

# #Fig.S5
```{r, fig.height=4, fig.width=8}
d.bx_red %>% 
  filter(trt == "W") %>% 
  ggplot(aes(x = PCA_soil_chem_1, y = Concentration, color = trt)) +
  geom_point(aes(x = PCA_soil_chem_1, y = Concentration, color = trt),
            show.legend = FALSE,  size = 2, alpha = 3/4) +
  stat_smooth(method = "lm", formula = 'y ~ x',
                color = "black", se = FALSE, fullrange = TRUE) +
  ylab("Concentration (ng/ml)") +
  xlab("Soil chemistry PC1") +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~")),
           colour = "dimgrey", vjust = 1.5) +
  scale_color_manual(name = "Genotype",
                     labels = c(W = "WT", b = expression(italic(bx1))),
                     values = c(W = "gold3", b = "darkgreen")) +
  facet_wrap(vars(Compound_name), nrow = 1, scales = "free")

ggsave(here("Figures", "individual_figures", "Fig.S5.svg"), 
       width = 34, height =  7, units = "cm")

```



# Fig.4A
```{r}
d.bx_long %>% 
  filter(trt == "W") %>% 
  mutate(sample_origin = factor(sample_origin),
         is_ampo = case_when( Compound_name == "AMPO" ~ "AMPO",
                             TRUE ~ "Benzoxazinoids"),
         is_ampo = factor(is_ampo, levels =
                               c("Benzoxazinoids", "AMPO"))) %>% 
  group_by(is_ampo, sample_origin, Compound_name) %>%
  summarise(
    Conc = mean(Concentration, na.rm = TRUE),
    n = sum(!is.na(Concentration)),
    sd = sd(Concentration, na.rm = TRUE),
    se = sd/sqrt(n),
    .groups = "drop") %>%
  ggplot(aes(sample_origin, Conc, color = Compound_name, group = Compound_name)) +
  geom_line() + 
  geom_point(size = 2.5) +
  geom_errorbar(aes(ymin = Conc - se, ymax = Conc + se), width = 0.2) +
  scale_y_continuous(limits = c(0, NA)) +
  scale_color_manual(name = "Compound",
                     values = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", 
                                "#0072B2", "#D55E00", "#CC79A7"))+ 
  scale_x_discrete(labels = c(Conditioning_phase = "Maize harvest",
                              Wheat_sowing = "Wheat sowing",
                              Wheat_end_veg_stage = "Wheat growth")) +
  facet_grid(vars(is_ampo), scales = "free") +
  labs(x = "Sampling timepoint",
       y = "Concentration (ng/ml)")

ggsave(here("Figures", "individual_figures", "Fig.4A.svg"), 
       width = 14, height =  14, units = "cm")

```



# Wheat start
```{r include=FALSE}
d.bx_w_start <- d.bx_long %>% 
  filter(sample_origin == "Wheat_sowing")
```

Fig.4B (top)
```{r warning=FALSE}
# test for differences between genotypes by Wilcoxon Rank Sum test and correct for multiple testing
p_values <- d.bx_w_start %>% 
  mutate(Concentration = case_when(
    is.na(Concentration) ~ 0,
    TRUE ~ Concentration))%>%
  group_by(Compound_name) %>% 
  nest() %>% 
  mutate( fit = map(data, ~ wilcox.test(Concentration ~ trt, data = ., exact = FALSE)),
          results = map(fit, broom::glance)) %>% 
  unnest(results)%>% 
  select(1, 5)%>%
  ungroup() %>% 
  mutate(
    n      = length(.$p.value),
    p.ad   = p.adjust(p.value, method = "fdr"),
    p.ad_c = case_when(
      p.ad < 0.001 ~ as.character("p < 0.001"),
      TRUE  ~ paste0("p = ", sprintf("%.3f", p.ad))),
    x      = 1.5,
    y      = Inf)

p.4b1 <- d.bx_w_start %>% 
  ggplot(aes(x = trt, y = Concentration)) +
  stat_summary(aes(fill = trt), fun = mean, geom = "bar", 
                 position = position_dodge(width = 0.8),
                 color = "black",  size = 0.7, alpha = 0.85, show.legend = FALSE) +
  stat_summary(fun.data = mean_se, geom = "errorbar", 
                 position = position_dodge(width = 0.8),
                 size = 0.8, color = "black", width = 0.2, alpha = 0.65) +
  geom_quasirandom(aes(color = PCA_soil_chem_1), size = 1.8, shape = 16, alpha = 1,
                     width = 0.2) +
  ylab("Concentration (ng/ml)") +
  xlab("Conditioning") +
  scale_y_continuous(expand = expansion(mult = c(0, 0.18))) +
  scale_x_discrete(labels = c(W = "WT", b = expression(italic(bx1)))) +
  scale_fill_manual(values = c(W = "gold2", b = "darkgreen")) +
  scale_color_continuous(name = "Soil chemistry PC1") +
  facet_wrap(vars(Compound_name), nrow = 1, scales = "free") + 
  geom_text(data = p_values, 
            mapping = aes(x = x, y = y, label = p.ad_c, fill = NULL),
            vjust = 2, size = 3.5) +
  theme(legend.position = "top", legend.key.height = unit(.3, "cm"),
        legend.margin = margin(b =  0, unit='cm'),
        legend.box.margin = margin(b = -10, unit = "pt"))

p.4b1

```

Fig.S9 (top)
```{r fig.width=14, fig.height=8, warning=FALSE}
p.S9a <- d.bx_w_start %>% 
  ggplot(aes(x = PCA_soil_chem_1, y = Concentration, color = trt)) +
  geom_point(size = 2, alpha = 3/4) +
  stat_smooth(method = "lm", formula = 'y ~ x', 
              se = FALSE, fullrange = TRUE) +
  ylab("Concentration (ng/ml)") +
  xlab("Soil chemistry PC1") +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~"))) +
  scale_color_manual(name = "Conditioning",
                     labels = c(W = "WT", b = expression(italic(bx1))),
                     values = c(W = "gold3", b = "darkgreen")) +
  facet_wrap(vars(Compound_name), nrow = 1, scales = "free") +
  theme(legend.position = "top")

p.S9a

```


# Wheat vegetative growth
```{r include=FALSE}
d.bx_w_growth <- d.bx_long %>% 
  filter(sample_origin == "Wheat_end_veg_stage")
```

Fig.4B (bottom)
```{r warning=FALSE}
# test for differences between genotypes by Wilcoxon Rank Sum test and correct for multiple testing
p_values <- d.bx_w_growth %>% 
  mutate(Concentration = case_when(
    is.na(Concentration) ~ 0,
    TRUE ~ Concentration))%>%
  group_by(Compound_name) %>% 
  nest() %>% 
  mutate( fit = map(data, ~ wilcox.test(Concentration ~ trt, data = ., exact = FALSE)),
          results = map(fit, broom::glance)) %>% 
  unnest(results)%>% 
  select(1, 5)%>%
  ungroup() %>% 
  mutate(
    n      = length(.$p.value),
    p.ad   = p.adjust(p.value, method = "fdr"),
    p.ad_c = case_when(
      p.ad < 0.001 ~ as.character("p < 0.001"),
      TRUE  ~ paste0("p = ", sprintf("%.3f", p.ad))),
    x      = 1.5,
    y      = Inf)


p.4b2 <- d.bx_w_growth %>% 
  ggplot(aes(x = trt, y = Concentration)) +
  stat_summary(aes(fill = trt), fun = mean, geom = "bar", 
                 position = position_dodge(width = 0.8),
                 color = "black",  size = 0.7, alpha = 0.85, show.legend = FALSE) +
  stat_summary(fun.data = mean_se, geom = "errorbar", 
                 position = position_dodge(width = 0.8),
                 size = 0.8, color = "black", width = 0.2, alpha = 0.65) +
  geom_quasirandom(aes(color = PCA_soil_chem_1), size = 1.8, shape = 16, 
                   alpha = 1, width = 0.2) +
  ylab("Concentration (ng/ml)") +
  xlab("Conditioning") +
  scale_y_continuous(expand = expansion(mult = c(0, 0.18))) +
  scale_x_discrete(labels = c(W = "WT", b = expression(italic(bx1)))) +
  scale_fill_manual(values = c(W = "gold2", b = "darkgreen")) +
  scale_color_continuous(name = "Soil chemistry PC1") +
  facet_wrap(vars(Compound_name), nrow = 1, scales = "free") + 
  geom_text(data = p_values, 
            mapping = aes(x = x, y = y, label = p.ad_c, fill = NULL),
            vjust = 2, size = 3.5) +
  theme(legend.position = "top", legend.key.height = unit(.3, "cm"))
```

# #Fig.4B
```{r}
# combine plots
p <- plot_grid(p.4b1 + theme(legend.position = "none"),
               p.4b2 + theme(legend.position = "none"), nrow = 2)

# add legend
legend <- get_legend(p.4b2)
fig.4b <- plot_grid(legend, p, nrow = 2, rel_heights = c(0.05, 1))
fig.4b

ggsave(here("Figures", "individual_figures", "Fig.4B.svg"), 
       width = 20, height =  14, units = "cm")
```

Fig.S9B (bottom)
```{r fig.width=14, fig.height=8, warning=FALSE}
p.S9b <-  d.bx_w_growth %>% 
  ggplot(aes(x = PCA_soil_chem_1, y = Concentration, color = trt)) +
  geom_point(size = 2, alpha = 3/4) +
  stat_smooth(method = "lm", formula = 'y ~ x', 
              se = FALSE, fullrange = TRUE) +
  ylab("Concentration (ng/ml)") +
  xlab("Soil chemistry PC1") +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~"))) +
  scale_color_manual(name = "Conditioning",
                     labels = c(W = "WT", b = expression(italic(bx1))),
                     values = c(W = "gold3", b = "darkgreen")) +
  facet_wrap(vars(Compound_name), nrow = 1, scales = "free") +
  theme(legend.position = "top")
```

# #Fig.S9
```{r}
# combine plots
p <- plot_grid(p.S9a + theme(legend.position = "none"),
               p.S9b + theme(legend.position = "none"), nrow = 2)

# add legend
legend <- get_legend(p.S9b)
p.S9 <- plot_grid(legend, p, nrow = 2, rel_heights = c(0.05, 1))
p.S9

ggsave(here("Figures", "individual_figures", "Fig.S9.svg"), 
       width = 30, height =  14, units = "cm")
```

## Version control
```{r}
sessionInfo()

```
