==> All_Other_Analysis/Scripts/03_BXs_growth_chamber.Rmd <==
---
title: "Experiment BX-feedback M18W19 <br> Effect of soil origin on maize BXs"
author: "Valentin Gfeller"
date: "`r Sys.Date()`"
output: 
  html_document: 
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
#global knitter option
knitr::opts_chunk$set(
	echo = FALSE,
	fig.align = "center",
	warning = FALSE,
	dev = "svglite"
)

# load packages
library(tidyverse); packageVersion("tidyverse")
library(ggbeeswarm); packageVersion("ggbeeswarm")
library(viridis); packageVersion("viridis")
library(car); packageVersion("car")
library(emmeans); packageVersion("emmeans")
library(here); packageVersion("here")
library(knitr); packageVersion("knitr")
library(ggpubr); packageVersion("ggpubr")
library(ggtext); packageVersion("ggtext")
library(cowplot); packageVersion("cowplot")

# source handy functions
source(here("Functions", "fun_vg.R"))

# set global option for ggplot2
theme_set(theme_bw(base_size = 11))

```

```{r warning=FALSE, include=FALSE}
# Import
d.ori_exu <- read.delim(here("Data", "05_BX_climate_chamber",
                             "Exp_m18w19_bx_exudates.csv"),
                        sep = ",", header = TRUE)

# convert to long format
d.fert_maize_long <- d.ori_exu %>%
  pivot_longer(cols = DIMBOA_Glc:MBOA,
               names_to = "Compound_name",
               values_to = "ng_mL")%>% 
  mutate(ng_cm = ng_mL/8, 
         Compound_name = factor(Compound_name, 
                                levels = c("DIMBOA_Glc", "HDMBOA_Glc",
                                           "DIMBOA", "HMBOA", "MBOA")))

str(d.fert_maize_long)

```

# #Fig.5B
```{r fig.height=4, fig.width=13, warning=FALSE}
d.fert_maize_long_orig <- d.fert_maize_long   

lab_name <- c(DIMBOA_Glc = "DIMBOA-Glc", HDMBOA_Glc = "HDMBOA-Glc",
              DIMBOA = "DIMBOA", HMBOA = "HMBOA", MBOA = "MBOA", AMPO = "AMPO")

# Test for differences between genotypes by t.test
p_values <- d.fert_maize_long_orig %>%
  group_by(Compound_name) %>% 
  nest() %>% 
  mutate( fit = map(data, ~ t.test(ng_cm ~ soil_origin, data = ., exact = FALSE)),
          results = map(fit, broom::glance)) %>% 
  unnest(results)%>% 
  select(1, 8)%>%
  ungroup() %>% 
  mutate(
    n      = length(.$p.value),
    p.ad   = p.adjust(p.value, method = "fdr"),
    p.ad_c = case_when(
      p.ad < 0.001 ~ as.character("p < 0.001"),
      TRUE  ~ paste0("p = ", sprintf("%.3f", p.ad))),
    x      = 1.5,
    y      = Inf)

d.fert_maize_long_orig %>%
  rbind(tibble(pot_id = NA, soil_origin = "South", dw_shoots = NA, dw_roots = NA, 
          Compound_name = "AMPO", ng_mL = NA, ng_cm = NA)) %>% 
  rbind(tibble(pot_id = NA, soil_origin = "North", dw_shoots = NA, dw_roots = NA, 
          Compound_name = "AMPO", ng_mL = NA, ng_cm = NA)) %>% 
  mutate(Compound_name = factor(Compound_name, 
                                levels = c("DIMBOA_Glc", "HDMBOA_Glc", "DIMBOA",
                                           "HMBOA", "MBOA", "AMPO"))) %>% 
  ggplot(aes(x = soil_origin, y = ng_cm, fill = soil_origin)) +
  geom_boxplot(outlier.shape = NA, alpha = 0.4, color = "grey",
               show.legend = FALSE) +
  stat_summary(fun = mean, geom = "point", shape = 18, 
               size = 5, color = "black", show.legend = FALSE) +
  stat_summary(fun.data = mean_se, geom = "errorbar", show.legend = FALSE,
               size = 0.9, color = "black", width = 0.2, alpha = 0.65) +
  geom_quasirandom(size = 2, shape = 16, alpha = 0.5, 
                   width = 0.2, dodge.width = 0.8, show.legend = FALSE) +
  scale_fill_manual(values = c(South = "#D95F02", North  = "#1B9E77")) +
  scale_x_discrete(name = "Soil origin") +
  ylab("Root exudation (ng/cm root tip)") +
  scale_y_continuous(limits = c(0, NA), expand = expansion(mult = c(0, 0.16))) +
  facet_wrap(vars(Compound_name), scales = "free", nrow = 1,
             labeller = labeller(Compound_name = lab_name)) + 
  geom_text(
      data    = p_values,
      mapping = aes(x = x, y = y, label = p.ad_c, fill = NULL),
      vjust   = 2,
      size    = 3.5)

ggsave(here("Figures", "individual_figures", "Fig.5B.svg"), 
       width = 10, height =  3)


# Sample size
d.fert_maize_long_orig %>% 
  group_by(Compound_name, soil_origin) %>% 
  summarise(
    n = sum(!is.na(ng_cm))
  )
```


## Soil BXs
```{r include=FALSE}
# Import
d.ori_soil <- read.delim(here("Data", "05_BX_climate_chamber",
                             "Exp_m18w19_bx_soil.csv"),
                        sep = ",", header = TRUE)

# convert to long format
d.fert_soil_maize_long <- d.ori_soil %>%
  pivot_longer(cols = DIMBOA_Glc:AMPO,
               names_to = "Compound_name",
               values_to = "ng_mL")%>%
  mutate(ng_mL_gRoot = ng_mL/dw_roots,
         Compound_name = factor(Compound_name, 
                                levels = c("DIMBOA_Glc", "HDMBOA_Glc", 
                                           "DIMBOA", "HMBOA", "MBOA", "AMPO")))

str(d.fert_soil_maize_long)

```

# #Fig.5C
```{r fig.height=4, fig.width=13, warning=FALSE}
lab_name <- c(DIMBOA_Glc = "DIMBOA-Glc", HDMBOA_Glc = "HDMBOA-Glc",
              DIMBOA = "DIMBOA", HMBOA = "HMBOA", MBOA = "MBOA", AMPO = "AMPO")

# Test for differences between genotypes by t.test
p_values <- d.fert_soil_maize_long %>%
  group_by(Compound_name) %>% 
  nest() %>% 
  mutate( fit = map(data, ~ t.test(ng_mL_gRoot ~ soil_origin, data = ., 
                                   exact = FALSE)),
          results = map(fit, broom::glance)) %>% 
  unnest(results)%>% 
  select(1, 8)%>%
  ungroup() %>% 
  mutate(
    n      = length(.$p.value),
    p.ad   = p.adjust(p.value, method = "fdr"),
    p.ad_c = case_when(
      p.ad < 0.001 ~ as.character("p < 0.001"),
      TRUE  ~ paste0("p = ", sprintf("%.3f", p.ad))),
    x      = 1.5,
    y      = Inf)

d.fert_soil_maize_long %>%
  ggplot(aes(x = soil_origin, y = ng_mL_gRoot, fill = soil_origin)) +
  geom_boxplot(outlier.shape = NA, alpha = 0.4, color = "grey",
               show.legend = FALSE) +
  stat_summary(fun = mean, geom = "point", shape = 18, 
               size = 5, color = "black", show.legend = FALSE) +
  stat_summary(fun.data = mean_se, geom = "errorbar", show.legend = FALSE,
               size = 0.9, color = "black", width = 0.2, alpha = 0.65) +
  geom_quasirandom(size = 2, shape = 16, alpha = 0.5, 
                   width = 0.2, dodge.width = 0.8, show.legend = FALSE) +
  scale_fill_manual(values = c(South = "#D95F02", North  = "#1B9E77")) +
  scale_x_discrete(name = "Soil origin") +
  ylab("Soil BXs (ng/mL*gDW)") +
  scale_y_continuous(limits = c(0, NA), expand = expansion(mult = c(0, 0.16))) +
  facet_wrap(vars(Compound_name), scales = "free", nrow = 1,
             labeller = labeller(Compound_name = lab_name)) + 
    geom_text(
      data    = p_values,
      mapping = aes(x = x, y = y, label = p.ad_c, fill = NULL),
      vjust   = 2,
      size    = 3.5)

ggsave(here("Figures", "individual_figures", "Fig.5C.svg"), 
       width = 10, height =  3)

# Sample size
d.fert_soil_maize_long %>% 
  group_by(Compound_name, soil_origin) %>% 
  summarise(
    n = sum(!is.na(ng_mL_gRoot))
  )

```


#Version control
```{r}
sessionInfo()
```
